COMP: cast class enum to int for array access
Static cast the MemorySpace class enum to an int when
accessing the umpire_type array.

+++ src/axom/core/tests/core_memory_management.cpp
@@ -38,7 +38,8 @@ void check_alloc_and_free( axom::MemorySpace spaceId )
 
   auto& rm = umpire::ResourceManager::getInstance();
   umpire::Allocator allocator = rm.getAllocator( data );
-  EXPECT_EQ( allocator.getId(), axom::internal::umpire_type[ spaceId ] );
+  EXPECT_EQ( allocator.getId(),
+             axom::internal::umpire_type[ static_cast< int >(spaceId) ] );
 
 #endif
